# api/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API to get Apple product prices.",
        "title": "Apple Product Prices API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/exchange": {
            "get": {
                "description": "Retrieves the latest exchange rates from an external API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange"
                ],
                "summary": "Get exchange rates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base currency (default is USD)",
                        "name": "base",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExchangeRatesResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch or decode data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Retrieves a single Apple product by its name from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the product",
                        "name": "product_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Product"
                        }
                    },
                    "400": {
                        "description": "Missing 'product_name' query parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/name": {
            "get": {
                "description": "Retrieves a list of all Apple product names, including images.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get all product names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ExchangeRatesResponse": {
            "type": "object",
            "properties": {
                "base_code": {
                    "type": "string"
                },
                "conversion_rates": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                },
                "result": {
                    "type": "string"
                },
                "time_last_update_unix": {
                    "type": "integer"
                },
                "time_last_update_utc": {
                    "type": "string"
                }
            }
        },
        "handlers.PriceEntry": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "vat": {
                    "type": "number"
                }
            }
        },
        "handlers.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/handlers.PriceEntry"
                        }
                    }
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                }
            }
        }
    }
}